+++ app/api/masterSize/controller.js
@@ -81,7 +81,7 @@ export const editSize = async (req, res) => {
             data: response
         }
 
-        return res.status(201).json({ umpanBalik })
+        return res.status(200).json({ umpanBalik })
     } catch (err) {
         const umpanBalik = {
             error: true,