+++ app/api/productPrice/controller.js
@@ -5,7 +5,7 @@ const prisma = new PrismaClient()
 
 let customUnix = Math.floor(new Date().getTime() / 1000.0)
 
-export const getProductPrice = async (req, res) => {
+export const getAllProductPrice = async (req, res) => {
     const { master_product_id } = req.params
     try {
         const response = await prisma.Product_Price.findMany({
@@ -34,6 +34,7 @@ export const createProductPrice = async (req, res) => {
     const {
         product_id,
         size_id,
+        is_promo,
         price,
         email,
     } = req.body
@@ -60,6 +61,7 @@ export const createProductPrice = async (req, res) => {
             uuid,
             product_id,
             size_id,
+            is_promo,
             price,
             post_by: email,
             custom_unix_createdAt: customUnix,
@@ -85,11 +87,39 @@ export const createProductPrice = async (req, res) => {
     }
 }
 
+export const editProductPrice = async (req, res) => {
+    const { uuid } = req.params
+    try {
+        const response = await prisma.$transaction([
+            prisma.Product_Price.findFirst({
+                where: {
+                    uuid: uuid,
+                    deleted: false
+                }
+            })
+        ])
+        const umpanBalik = {
+            error: false,
+            message: 'success',
+            data: response
+        }
+        res.status(200).json({ umpanBalik })
+    } catch (err) {
+        const umpanBalik = {
+            error: true,
+            message: err.message,
+            data: 'kosong'
+        }
+        return res.status(500).json({ umpanBalik: umpanBalik || `Internal server error` })
+    }
+}
+
 export const updateProductPrice = async (req, res) => {
-    const { master_product_id, product_price_id } = req.params
+    const { product_price_id } = req.params
     const {
         product_id,
         size_id,
+        is_promo,
         price,
         email,
     } = req.body
@@ -98,6 +128,7 @@ export const updateProductPrice = async (req, res) => {
         const formData = {
             product_id,
             size_id,
+            is_promo,
             price,
             edited_by: email,
             custom_unix_updatedAt: customUnix
@@ -106,7 +137,6 @@ export const updateProductPrice = async (req, res) => {
             prisma.Product_Price.update({
                 where: {
                     uuid: product_price_id,
-                    product_id: master_product_id
                 },
                 data: formData
             })
@@ -129,8 +159,7 @@ export const updateProductPrice = async (req, res) => {
 }
 
 export const deleteProductPrice = async (req, res) => {
-    const { master_product_id, product_price_id } = req.params
-    const { email } = req.body
+    const { user_id, product_price_id } = req.params
     try {
         prisma.$use(async (params, next) => {
             // Check incoming query type
@@ -144,21 +173,19 @@ export const deleteProductPrice = async (req, res) => {
         })
 
         const formData = {
-            deleted_by: email,
+            deleted_by: user_id,
             custom_unix_soft_delete: customUnix
         }
         const response = await prisma.$transaction([
             prisma.Product_Price.update({
                 where: {
                     uuid: product_price_id,
-                    product_id: master_product_id
                 },
                 data: formData
             }),
             prisma.Product_Price.delete({
                 where: {
                     uuid: product_price_id,
-                    product_id: master_product_id
                 }
             })
         ])